Current/pyproject.toml
[project]
name = "leanknowledge"
version = "0.2.0"
description = "LLM-powered formalization of mathematical theorems into verified Lean 4 code"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.0",
    "pymupdf>=1.24",
    "litellm>=1.0",
]

[project.optional-dependencies]
google = ["google-cloud-documentai>=2.20"]
test = ["pytest>=8.0"]

[project.scripts]
leanknowledge = "leanknowledge.pipeline:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/leanknowledge"]
